convert form to vue component
Form takes in errors as props then displays on correct
form input. The datepicker doesn't turn the field red
becuase there's no way to bind the class to the custom
class prop it uses.

New forms now have to render on error instead of redirect
so that the errors variable is available.